webcaster/requirements.txt
Flask
confluent-kafka
python-dotenv
Flask-CORS